Add eslint for single-file components
Note that the linter acts on Vue components in both `.js` and `.vue`
files.

To keep the changes minimal and targeted, for now the linter uses a very
permissive rule-set which essentially means that the linter is mostly
disabled. As a separate change, we should fix all the lint issues and
tighten the rule-set to `plugin:vue/recommended`.