Move swiper initialization to webpack bundle
Currently most of doccano's Javascript is in webpack bundles, except for
a few snippets like swiper. To increase consistency, this change moves
the swiper code also into a webpack bundle.

In addition to increasing consistency, this also has the advantages of:

1) Giving developers a single location/directory to look for all the
   Javascript.

2) Making the swiper dependency explicit by listing it in the
   package.json manifest

3) Ensuring that swiper gets built and minified in the same way as other
   dependencies.

4) Making swiper available when developing offline.

In the future, it would also be worth considering to move all the CSS
includes into webpack.