Organize webpack entrypoints
Since the move to Vue single-file components, the number of webpack
entrypoints has greatly increased and the files somewhat clutter up the
frontend folder. As such, to increase discoverability and organization
of the code, this change moves all the webpack entrypoints into a single
folder so that the files can more easily be ignored when not adding or
changing an entrypoint.