Proxy all requests at /v1 server-side to the API
This way all calls to the API are done through the frontend, and the API
is not required to be exposed directly to browsers.

Additionally, it's now possible to specify the url of the API using an
environment variable. Downside is this needs to be done explicitly.

Finally, the container is able to execute in a bare-bones configuration.